// thunderstorm/frontend/src/main/modules/routing/types.ts
import * as React from 'react';
import {QueryParams} from '../../shared.js';
import {ResolvableContent} from '@nu-art/ts-common';
import {ModuleFE_BaseDB} from '../db-api-gen/ModuleFE_BaseDB.js';
import {AwaitModule_LoaderProps} from '../../components/AwaitModules/AwaitModules.js';

export type TS_Route<T extends QueryParams = QueryParams> = {
	key: string;
	path: string;
	enabled?: () => boolean
	element?: React.ReactNode;
	Component?: React.ComponentType<any>;
	paramKeys?: (keyof T)[]
	fallback?: boolean;
	index?: boolean;
	children?: TS_Route<any>[]
	modulesToAwait?: ResolvableContent<(ModuleFE_BaseDB<any, any>)[]>;
	awaitModulesLoader?: ResolvableContent<React.ReactNode, [AwaitModule_LoaderProps]>;
	awaitSync?: boolean;
	awaitSyncLoader?: ResolvableContent<React.ReactNode>;
}